pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-trader"
version = "0.1.0"
description = "Polymarket Multi-LLM Paper Trading System"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.30,<1.0",
    "alembic>=1.14,<2.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "httpx>=0.28,<1.0",
    "websockets>=14.0,<15.0",
    "apscheduler>=3.10,<4.0",
    "structlog>=24.0,<26.0",
    "pyyaml>=6.0,<7.0",
    "feedparser>=6.0,<7.0",
    "beautifulsoup4>=4.12,<5.0",
    "openai>=1.60,<3.0",
    "anthropic>=0.40,<1.0",
    "google-genai>=1.0,<2.0",
    "pytrends>=4.9,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=6.0,<7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]
